Initial batch of IU-specific edits.
After forking the repo, this commit represents the first set of changes
we had implemented prior to officially forking the repository. This
will allow us to merge Ken’s updates and changes with our own,
retaining a relationship between the repositories. See the
readme-oidd.txt file for summary of IU changes.